+++ __SITE__/analyse/_lib/_required/user/instance.rb
@@ -6,9 +6,4 @@ class User
     @is_analyste
   end
 
-  # Retourne true si l'user a fait une demande de participation
-  def postulant?
-    @is_postulant.nil? && @is_postulant = identified? && (data[:options][16].to_i == 1)
-    @is_postulant
-  end
 end #/User