+++ __SITE__/analyse/lire/main.rb
@@ -1,4 +1,1 @@
 # encoding: utf-8
-def film
-  @film ||= Analyse::Film.new(site.route.objet_id)
-end