+++ __SITE__/analyse/lire/partial/lecon/main.rb
@@ -1,0 +1,16 @@
+# encoding: utf-8
+class Analyse
+  class Film
+    class << self
+
+      # Retourne la liste des tous les films qui font l'objet
+      # d'une lecon tirée d'un film
+      def films_with_lecon
+        @films_with_lecon ||= 
+          begin
+          end
+      end
+
+    end # << self
+  end #/Film
+end #/Analyse